Separate schema dump functionality into smaller files
