+++ lib/data_migrate/data_migrator.rb
@@ -94,93 +94,6 @@ module DataMigrate
       def db_config
         ActiveRecord::Base.configurations[Rails.env || 'development'] || ENV["DATABASE_URL"]
       end
-
     end
-  end
-
-  ##
-  # This class extends DatabaseTasks to override the schema_file method.
-  class DatabaseTasks
-    extend ActiveRecord::Tasks::DatabaseTasks
-
-    def self.schema_file(format = ActiveRecord::Base.schema_format)
-      case format
-      when :ruby
-        File.join(db_dir, "data_schema.rb")
-      else
-        message = "Only Ruby-based data_schema files are supported at this time."
-        Kernel.abort message
-      end
-    end
-  end
-
-  ##
-  # Provides the definition method for data_schema.rb
-  class Data < ActiveRecord::Schema
-    # This method is based on the following two methods
-    #   ActiveRecord::Schema#define
-    #   ActiveRecord::ConnectionAdapters::SchemaStatements#ssume_migrated_upto_version
-    def define(info)
-      DataMigrate::DataMigrator.assure_data_schema_table
-
-      return if info[:version].blank?
-
-      version = info[:version].to_i
-      sm_table = quote_table_name(DataMigrate::DataMigrator.schema_migrations_table_name)
-      migrated = select_values("SELECT version FROM #{sm_table}").map(&:to_i)
-
-      versions = []
-      Dir.foreach(DataMigrate::DataMigrator.full_migrations_path) do |file|
-        match_data = DataMigrate::DataMigrator.match(file)
-        versions << match_data[1].to_i if match_data
-      end
-
-      unless migrated.include?(version)
-        execute "INSERT INTO #{sm_table} (version) VALUES ('#{version}')"
-      end
-
-      inserted = Set.new
-      (versions - migrated).each do |v|
-        if inserted.include?(v)
-          raise "Duplicate data migration #{v}. Please renumber your data migrations to resolve the conflict."
-        elsif v < version
-          execute "INSERT INTO #{sm_table} (version) VALUES ('#{v}')"
-          inserted << v
-        end
-      end
-    end
-  end
-
-  ##
-  # Provides the capability to write the current data schema version to the data_schema file
-  # Based on ActiveRecord::SchemaDumper
-  class SchemaDumper
-    private_class_method :new
-
-    class << self
-      def dump(connection = ActiveRecord::Base.connection, stream = STDOUT)
-        new(connection).dump(stream)
-        stream
-      end
-    end
-
-    def dump(stream)
-      define_params = @version ? "version: #{@version}" : ""
-
-      if stream.respond_to?(:external_encoding) && stream.external_encoding
-        stream.puts "# encoding: #{stream.external_encoding.name}"
-      end
-
-      stream.puts "DataMigrate::Data.define(#{define_params})"
-
-      stream
-    end
-
-    private
-
-      def initialize(connection)
-        @connection = connection
-        @version = DataMigrate::DataSchemaMigration.all.map { |x| x.version.to_i }.max rescue 0
-      end
   end
 end
